Database.Beam

Description

Top-level Beam module. This module re-exports all the symbols necessary for most common user operations.

The most interesting modules are Database.Beam.Schema and Database.Beam.Query.

This is mainly reference documentation. Most users will want to consult the manual.

Documentation

module Database.Beam.Query

module Database.Beam.Schema

class (BeamBackend be, Monad m) => MonadBeam be m | m -> be where

A class that ties together a monad with a particular backend

Provided here is a low-level interface for executing commands. The 'run*' functions are wrapped by the appropriate functions in Query.

This interface is very high-level and isn't meant to expose the full power of the underlying database. Namely, it only supports simple data retrieval strategies. More complicated strategies (for example, Postgres's COPY) are supported in individual backends. See the documentation of those backends for more details.

Minimal complete definition

runReturningMany

Methods

runReturningMany

Arguments

:: FromBackendRow be x

=> BeamSqlBackendSyntax be
The query to run

-> (m (Maybe x) -> m a)
Reader action that will be called with a function to fetch the next row

-> m a

Run a query determined by the given syntax, providing an action that will be called to consume the results from the database (if any). The action will get a reader action that can be used to fetch the next row. When this reader action returns Nothing, there are no rows left to consume. When the reader action returns, the database result is freed.

runNoReturn :: BeamSqlBackendSyntax be -> m ()

Run the given command and don't consume any results. Useful for DML statements like INSERT, UPDATE, and DELETE, or DDL statements.

runReturningOne :: FromBackendRow be x => BeamSqlBackendSyntax be -> m (Maybe x)

Run the given command and fetch the unique result. The result is Nothing if either no results are returned or more than one result is returned.

runReturningFirst :: FromBackendRow be x => BeamSqlBackendSyntax be -> m (Maybe x)

Run the given command and fetch the first result. The result is Nothing if no results are returned. This is not guaranteed to automatically limit the query to one result.

runReturningList :: FromBackendRow be x => BeamSqlBackendSyntax be -> m [x]

Run the given command, collect all the results, and return them as a list. May be more convenient than runReturningMany, but reads the entire result set into memory.

class BeamBackend be => FromBackendRow be a where

Minimal complete definition

Nothing

Methods

fromBackendRow :: FromBackendRowM be a

Parses a beam row. This should not fail, except in the case of an internal bug in beam deserialization code. If it does fail, this should throw a BeamRowParseError.

default fromBackendRow :: (Typeable a, BackendFromField be a) => FromBackendRowM be a

valuesNeeded :: Proxy be -> Proxy a -> Int

Re-exports

class Monad m => MonadIO (m :: Type -> Type) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Minimal complete definition

from, to

data Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0