Database.Beam.Backend.SQL.Builder

Description

Provides a syntax SqlSyntaxBuilder that uses a Builder to construct SQL expressions as strings. Mainly serves as documentation for how to write a syntax for backends. Note that, although you can use this to turn most Q and QGenExprs into ByteString queries, it is very unwise to ship these to the database. This module does not take into account server-specific quoting. Some backends are very particular to quoting, and shipping arbitrary ByteStrings as queries can expose you to SQL injection vulnerabilities. Always use the provided backends to submit queries and data manipulation commands to the database.

Documentation

newtype SqlSyntaxBuilder

The main syntax. A wrapper over Builder

Constructors

SqlSyntaxBuilder

Fields

buildSql :: Builder

buildSepBy :: Builder -> [Builder] -> Builder

quoteSql :: Text -> Builder

renderSql :: SqlSyntaxBuilder -> String