Database.Beam.Backend.SQL.Row

Documentation

data FromBackendRowF be f where

Constructors

ParseOneField :: (BackendFromField be a, Typeable a) => (a -> f) -> FromBackendRowF be f

Alt :: FromBackendRowM be a -> FromBackendRowM be a -> (a -> f) -> FromBackendRowF be f

FailParseWith :: BeamRowReadError -> FromBackendRowF be f

newtype FromBackendRowM be a

Constructors

FromBackendRowM (F (FromBackendRowF be) a)

parseOneField :: (BackendFromField be a, Typeable a) => FromBackendRowM be a

peekField :: (Typeable a, BackendFromField be a) => FromBackendRowM be (Maybe a)

data ColumnParseError

The exact error encountered

Constructors

ColumnUnexpectedNull

ColumnNotEnoughColumns !Int

ColumnTypeMismatch

Fields

ctmHaskellType :: String

ctmSQLType :: String

ctmMessage :: String

ColumnErrorInternal String

data BeamRowReadError

An error that may occur when parsing a row. Contains an optional annotation of which column was being parsed (if available).

Constructors

BeamRowReadError

Fields

brreColumn :: !(Maybe Int)

brreError :: !ColumnParseError

class BeamBackend be => FromBackendRow be a where

Minimal complete definition

Nothing

Methods

fromBackendRow :: FromBackendRowM be a

Parses a beam row. This should not fail, except in the case of an internal bug in beam deserialization code. If it does fail, this should throw a BeamRowParseError.

default fromBackendRow :: (Typeable a, BackendFromField be a) => FromBackendRowM be a

valuesNeeded :: Proxy be -> Proxy a -> Int

Orphan instances

Generic (a, b, c, d, e, f, g, h)

Instance details

Associated Types

type Rep (a, b, c, d, e, f, g, h) :: Type -> Type

Methods

from :: (a, b, c, d, e, f, g, h) -> Rep (a, b, c, d, e, f, g, h) x

to :: Rep (a, b, c, d, e, f, g, h) x -> (a, b, c, d, e, f, g, h)