Database.Beam.Backend.SQL.SQL92

Description

Finally tagless encoding of SQL92 syntax

Finally tagless style

class HasSqlValueSyntax expr ty where

Methods

sqlValueSyntax :: ty -> expr

autoSqlValueSyntax :: (HasSqlValueSyntax expr String, Show a) => a -> expr

type Sql92SelectExpressionSyntax select = Sql92SelectTableExpressionSyntax (Sql92SelectSelectTableSyntax select)

type Sql92SelectProjectionSyntax select = Sql92SelectTableProjectionSyntax (Sql92SelectSelectTableSyntax select)

type Sql92SelectGroupingSyntax select = Sql92SelectTableGroupingSyntax (Sql92SelectSelectTableSyntax select)

type Sql92SelectFromSyntax select = Sql92SelectTableFromSyntax (Sql92SelectSelectTableSyntax select)

type Sql92InsertExpressionSyntax select = Sql92InsertValuesExpressionSyntax (Sql92InsertValuesSyntax select)

type Sql92TableNameSyntax select = Sql92TableSourceTableNameSyntax (Sql92FromTableSourceSyntax (Sql92SelectFromSyntax select))

type Sql92ValueSyntax cmdSyntax = Sql92ExpressionValueSyntax (Sql92ExpressionSyntax cmdSyntax)

type Sql92ExpressionSyntax cmdSyntax = Sql92SelectExpressionSyntax (Sql92SelectSyntax cmdSyntax)

type Sql92ExtractFieldSyntax cmdSyntax = Sql92ExpressionExtractFieldSyntax (Sql92ExpressionSyntax cmdSyntax)

type Sql92HasValueSyntax cmdSyntax = HasSqlValueSyntax (Sql92ValueSyntax cmdSyntax)

type Sql92SelectSanityCheck select = (Sql92FromExpressionSyntax (Sql92SelectTableFromSyntax (Sql92SelectSelectTableSyntax select)) ~ Sql92SelectTableExpressionSyntax (Sql92SelectSelectTableSyntax select), Sql92TableSourceSelectSyntax (Sql92FromTableSourceSyntax (Sql92SelectTableFromSyntax (Sql92SelectSelectTableSyntax select))) ~ select, Sql92ProjectionExpressionSyntax (Sql92SelectTableProjectionSyntax (Sql92SelectSelectTableSyntax select)) ~ Sql92SelectTableExpressionSyntax (Sql92SelectSelectTableSyntax select), Sql92OrderingExpressionSyntax (Sql92SelectOrderingSyntax select) ~ Sql92SelectTableExpressionSyntax (Sql92SelectSelectTableSyntax select), Sql92TableSourceExpressionSyntax (Sql92FromTableSourceSyntax (Sql92SelectTableFromSyntax (Sql92SelectSelectTableSyntax select))) ~ Sql92SelectTableExpressionSyntax (Sql92SelectSelectTableSyntax select))

type Sql92SanityCheck cmd = (Sql92SelectSanityCheck (Sql92SelectSyntax cmd), Sql92ExpressionValueSyntax (Sql92InsertValuesExpressionSyntax (Sql92InsertValuesSyntax (Sql92InsertSyntax cmd))) ~ Sql92ValueSyntax cmd, Sql92ExpressionValueSyntax (Sql92UpdateExpressionSyntax (Sql92UpdateSyntax cmd)) ~ Sql92ValueSyntax cmd, Sql92ExpressionValueSyntax (Sql92DeleteExpressionSyntax (Sql92DeleteSyntax cmd)) ~ Sql92ValueSyntax cmd, Sql92SelectTableExpressionSyntax (Sql92SelectSelectTableSyntax (Sql92SelectSyntax cmd)) ~ Sql92InsertValuesExpressionSyntax (Sql92InsertValuesSyntax (Sql92InsertSyntax cmd)), Sql92SelectTableExpressionSyntax (Sql92SelectSelectTableSyntax (Sql92SelectSyntax cmd)) ~ Sql92UpdateExpressionSyntax (Sql92UpdateSyntax cmd), Sql92DeleteExpressionSyntax (Sql92DeleteSyntax cmd) ~ Sql92UpdateExpressionSyntax (Sql92UpdateSyntax cmd), Sql92ExpressionSelectSyntax (Sql92InsertExpressionSyntax (Sql92InsertSyntax cmd)) ~ Sql92SelectSyntax cmd, Sql92InsertValuesSelectSyntax (Sql92InsertValuesSyntax (Sql92InsertSyntax cmd)) ~ Sql92SelectSyntax cmd, Sql92UpdateFieldNameSyntax (Sql92UpdateSyntax cmd) ~ Sql92ExpressionFieldNameSyntax (Sql92InsertValuesExpressionSyntax (Sql92InsertValuesSyntax (Sql92InsertSyntax cmd))))

type Sql92ReasonableMarshaller be = (FromBackendRow be SqlNull, FromBackendRow be Text, FromBackendRow be Bool, FromBackendRow be Char, FromBackendRow be Int16, FromBackendRow be Int32, FromBackendRow be Int64, FromBackendRow be LocalTime)

class Sql92DisplaySyntax syntax where

Type classes for syntaxes which can be displayed

Methods

displaySyntax :: syntax -> String

Render the syntax as a String, representing the SQL expression it stands for

class (IsSql92SelectSyntax (Sql92SelectSyntax cmd), IsSql92InsertSyntax (Sql92InsertSyntax cmd), IsSql92UpdateSyntax (Sql92UpdateSyntax cmd), IsSql92DeleteSyntax (Sql92DeleteSyntax cmd)) => IsSql92Syntax cmd where

Associated Types

type Sql92SelectSyntax cmd :: Type

type Sql92InsertSyntax cmd :: Type

type Sql92UpdateSyntax cmd :: Type

type Sql92DeleteSyntax cmd :: Type

Methods

selectCmd :: Sql92SelectSyntax cmd -> cmd

insertCmd :: Sql92InsertSyntax cmd -> cmd

updateCmd :: Sql92UpdateSyntax cmd -> cmd

deleteCmd :: Sql92DeleteSyntax cmd -> cmd

class (IsSql92SelectTableSyntax (Sql92SelectSelectTableSyntax select), IsSql92OrderingSyntax (Sql92SelectOrderingSyntax select)) => IsSql92SelectSyntax select where

Associated Types

type Sql92SelectSelectTableSyntax select :: Type

type Sql92SelectOrderingSyntax select :: Type

Methods

selectStmt

Arguments

:: Sql92SelectSelectTableSyntax select

-> [Sql92SelectOrderingSyntax select]

-> Maybe Integer
LIMIT

-> Maybe Integer
OFFSET

-> select

class (IsSql92ExpressionSyntax (Sql92SelectTableExpressionSyntax select), IsSql92AggregationExpressionSyntax (Sql92SelectTableExpressionSyntax select), IsSql92ProjectionSyntax (Sql92SelectTableProjectionSyntax select), IsSql92FromSyntax (Sql92SelectTableFromSyntax select), IsSql92GroupingSyntax (Sql92SelectTableGroupingSyntax select), IsSql92AggregationSetQuantifierSyntax (Sql92SelectTableSetQuantifierSyntax select), Sql92GroupingExpressionSyntax (Sql92SelectTableGroupingSyntax select) ~ Sql92SelectTableExpressionSyntax select, Sql92FromExpressionSyntax (Sql92SelectTableFromSyntax select) ~ Sql92SelectTableExpressionSyntax select, Sql92SelectSelectTableSyntax (Sql92SelectTableSelectSyntax select) ~ select, Eq (Sql92SelectTableExpressionSyntax select)) => IsSql92SelectTableSyntax select where

Associated Types

type Sql92SelectTableSelectSyntax select :: Type

type Sql92SelectTableExpressionSyntax select :: Type

type Sql92SelectTableProjectionSyntax select :: Type

type Sql92SelectTableFromSyntax select :: Type

type Sql92SelectTableGroupingSyntax select :: Type

type Sql92SelectTableSetQuantifierSyntax select :: Type

Methods

selectTableStmt

Arguments

:: Maybe (Sql92SelectTableSetQuantifierSyntax select)

-> Sql92SelectTableProjectionSyntax select

-> Maybe (Sql92SelectTableFromSyntax select)

-> Maybe (Sql92SelectTableExpressionSyntax select)
Where clause

-> Maybe (Sql92SelectTableGroupingSyntax select)

-> Maybe (Sql92SelectTableExpressionSyntax select)
having clause

-> select

unionTables :: Bool -> select -> select -> select

intersectTables :: Bool -> select -> select -> select

exceptTable :: Bool -> select -> select -> select

class (IsSql92InsertValuesSyntax (Sql92InsertValuesSyntax insert), IsSql92TableNameSyntax (Sql92InsertTableNameSyntax insert)) => IsSql92InsertSyntax insert where

Associated Types

type Sql92InsertValuesSyntax insert :: Type

type Sql92InsertTableNameSyntax insert :: Type

Methods

insertStmt

Arguments

:: Sql92InsertTableNameSyntax insert

-> [Text]
Fields

-> Sql92InsertValuesSyntax insert

-> insert

class IsSql92ExpressionSyntax (Sql92InsertValuesExpressionSyntax insertValues) => IsSql92InsertValuesSyntax insertValues where

Associated Types

type Sql92InsertValuesExpressionSyntax insertValues :: Type

type Sql92InsertValuesSelectSyntax insertValues :: Type

Methods

insertSqlExpressions :: [[Sql92InsertValuesExpressionSyntax insertValues]] -> insertValues

insertFromSql :: Sql92InsertValuesSelectSyntax insertValues -> insertValues

class (IsSql92ExpressionSyntax (Sql92UpdateExpressionSyntax update), IsSql92FieldNameSyntax (Sql92UpdateFieldNameSyntax update), IsSql92TableNameSyntax (Sql92UpdateTableNameSyntax update)) => IsSql92UpdateSyntax update where

Associated Types

type Sql92UpdateTableNameSyntax update :: Type

type Sql92UpdateFieldNameSyntax update :: Type

type Sql92UpdateExpressionSyntax update :: Type

Methods

updateStmt

Arguments

:: Sql92UpdateTableNameSyntax update

-> [(Sql92UpdateFieldNameSyntax update, Sql92UpdateExpressionSyntax update)]

-> Maybe (Sql92UpdateExpressionSyntax update)
WHERE

-> update

class (IsSql92TableNameSyntax (Sql92DeleteTableNameSyntax delete), IsSql92ExpressionSyntax (Sql92DeleteExpressionSyntax delete)) => IsSql92DeleteSyntax delete where

Minimal complete definition

deleteStmt

Associated Types

type Sql92DeleteTableNameSyntax delete :: Type

type Sql92DeleteExpressionSyntax delete :: Type

Methods

deleteStmt :: Sql92DeleteTableNameSyntax delete -> Maybe Text -> Maybe (Sql92DeleteExpressionSyntax delete) -> delete

deleteSupportsAlias :: Proxy delete -> Bool

Whether or not the DELETE command supports aliases

class IsSql92FieldNameSyntax fn where

Methods

qualifiedField :: Text -> Text -> fn

unqualifiedField :: Text -> fn

class IsSql92QuantifierSyntax quantifier where

Methods

quantifyOverAll :: quantifier

quantifyOverAny :: quantifier

class IsSql92ExtractFieldSyntax extractField where

Methods

secondsField :: extractField

minutesField :: extractField

hourField :: extractField

dayField :: extractField

monthField :: extractField

yearField :: extractField

class IsSql92DataTypeSyntax dataType where

Methods

domainType :: Text -> dataType

charType :: Maybe Word -> Maybe Text -> dataType

varCharType :: Maybe Word -> Maybe Text -> dataType

nationalCharType :: Maybe Word -> dataType

nationalVarCharType :: Maybe Word -> dataType

bitType :: Maybe Word -> dataType

varBitType :: Maybe Word -> dataType

numericType :: Maybe (Word, Maybe Word) -> dataType

decimalType :: Maybe (Word, Maybe Word) -> dataType

intType :: dataType

smallIntType :: dataType

floatType :: Maybe Word -> dataType

doubleType :: dataType

realType :: dataType

dateType :: dataType

timeType

Arguments

:: Maybe Word

-> Bool
With time zone

-> dataType

timestampType

Arguments

:: Maybe Word

-> Bool
With time zone

-> dataType

class (HasSqlValueSyntax (Sql92ExpressionValueSyntax expr) Int32, HasSqlValueSyntax (Sql92ExpressionValueSyntax expr) Bool, IsSql92FieldNameSyntax (Sql92ExpressionFieldNameSyntax expr), IsSql92QuantifierSyntax (Sql92ExpressionQuantifierSyntax expr), IsSql92DataTypeSyntax (Sql92ExpressionCastTargetSyntax expr), IsSql92ExtractFieldSyntax (Sql92ExpressionExtractFieldSyntax expr), Typeable expr) => IsSql92ExpressionSyntax expr where

Minimal complete definition

valueE, rowE, coalesceE, caseE, fieldE, andE, orE, addE, subE, mulE, divE, likeE, modE, overlapsE, nullIfE, positionE, eqE, neqE, ltE, gtE, leE, geE, castE, notE, negateE, isNullE, isNotNullE, isTrueE, isNotTrueE, isFalseE, isNotFalseE, isUnknownE, isNotUnknownE, charLengthE, octetLengthE, bitLengthE, lowerE, upperE, trimE, absE, extractE, existsE, uniqueE, subqueryE, currentTimestampE, defaultE, inE, inSelectE

Associated Types

type Sql92ExpressionQuantifierSyntax expr :: Type

type Sql92ExpressionValueSyntax expr :: Type

type Sql92ExpressionSelectSyntax expr :: Type

type Sql92ExpressionFieldNameSyntax expr :: Type

type Sql92ExpressionCastTargetSyntax expr :: Type

type Sql92ExpressionExtractFieldSyntax expr :: Type

Methods

valueE :: Sql92ExpressionValueSyntax expr -> expr

rowE :: [expr] -> expr

quantifierListE :: [expr] -> expr

coalesceE :: [expr] -> expr

caseE :: [(expr, expr)] -> expr -> expr

fieldE :: Sql92ExpressionFieldNameSyntax expr -> expr

betweenE :: expr -> expr -> expr -> expr

andE :: expr -> expr -> expr

orE :: expr -> expr -> expr

addE :: expr -> expr -> expr

subE :: expr -> expr -> expr

mulE :: expr -> expr -> expr

divE :: expr -> expr -> expr

likeE :: expr -> expr -> expr

modE :: expr -> expr -> expr

overlapsE :: expr -> expr -> expr

nullIfE :: expr -> expr -> expr

positionE :: expr -> expr -> expr

eqE :: Maybe (Sql92ExpressionQuantifierSyntax expr) -> expr -> expr -> expr

neqE :: Maybe (Sql92ExpressionQuantifierSyntax expr) -> expr -> expr -> expr

ltE :: Maybe (Sql92ExpressionQuantifierSyntax expr) -> expr -> expr -> expr

gtE :: Maybe (Sql92ExpressionQuantifierSyntax expr) -> expr -> expr -> expr

leE :: Maybe (Sql92ExpressionQuantifierSyntax expr) -> expr -> expr -> expr

geE :: Maybe (Sql92ExpressionQuantifierSyntax expr) -> expr -> expr -> expr

eqMaybeE :: expr -> expr -> expr -> expr

Compare the first and second argument for nullable equality, if they are both not null, return the result of the third expression

Some backends, like beam-postgres totally ignore the third result, because all equality there is sensible.

neqMaybeE :: expr -> expr -> expr -> expr

Compare the first and second argument for nullable equality, if they are both not null, return the result of the third expression

Some backends, like beam-postgres totally ignore the third result, because all equality there is sensible.

castE :: expr -> Sql92ExpressionCastTargetSyntax expr -> expr

notE :: expr -> expr

negateE :: expr -> expr

isNullE :: expr -> expr

isNotNullE :: expr -> expr

isTrueE :: expr -> expr

isNotTrueE :: expr -> expr

isFalseE :: expr -> expr

isNotFalseE :: expr -> expr

isUnknownE :: expr -> expr

isNotUnknownE :: expr -> expr

charLengthE :: expr -> expr

octetLengthE :: expr -> expr

bitLengthE :: expr -> expr

lowerE :: expr -> expr

upperE :: expr -> expr

trimE :: expr -> expr

absE :: expr -> expr

Included so that we can easily write a Num instance, but not defined in SQL92. Implementations that do not support this, should use CASE .. WHEN ..

extractE :: Sql92ExpressionExtractFieldSyntax expr -> expr -> expr

existsE :: Sql92ExpressionSelectSyntax expr -> expr

uniqueE :: Sql92ExpressionSelectSyntax expr -> expr

subqueryE :: Sql92ExpressionSelectSyntax expr -> expr

currentTimestampE :: expr

defaultE :: expr

inE :: expr -> [expr] -> expr

inSelectE :: expr -> Sql92ExpressionSelectSyntax expr -> expr

class IsSql92AggregationSetQuantifierSyntax (Sql92AggregationSetQuantifierSyntax expr) => IsSql92AggregationExpressionSyntax expr where

Associated Types

type Sql92AggregationSetQuantifierSyntax expr :: Type

Methods

countAllE :: expr

countE :: Maybe (Sql92AggregationSetQuantifierSyntax expr) -> expr -> expr

avgE :: Maybe (Sql92AggregationSetQuantifierSyntax expr) -> expr -> expr

maxE :: Maybe (Sql92AggregationSetQuantifierSyntax expr) -> expr -> expr

minE :: Maybe (Sql92AggregationSetQuantifierSyntax expr) -> expr -> expr

sumE :: Maybe (Sql92AggregationSetQuantifierSyntax expr) -> expr -> expr

class IsSql92AggregationSetQuantifierSyntax q where

Methods

setQuantifierDistinct :: q

setQuantifierAll :: q

class IsSql92ExpressionSyntax (Sql92ProjectionExpressionSyntax proj) => IsSql92ProjectionSyntax proj where

Associated Types

type Sql92ProjectionExpressionSyntax proj :: Type

Methods

projExprs :: [(Sql92ProjectionExpressionSyntax proj, Maybe Text)] -> proj

class IsSql92OrderingSyntax ord where

Associated Types

type Sql92OrderingExpressionSyntax ord :: Type

Methods

ascOrdering :: Sql92OrderingExpressionSyntax ord -> ord

descOrdering :: Sql92OrderingExpressionSyntax ord -> ord

class IsSql92TableNameSyntax tblName where

Methods

tableName

Arguments

:: Maybe Text
Schema

-> Text
Table name

-> tblName

class IsSql92TableNameSyntax (Sql92TableSourceTableNameSyntax tblSource) => IsSql92TableSourceSyntax tblSource where

Associated Types

type Sql92TableSourceSelectSyntax tblSource :: Type

type Sql92TableSourceExpressionSyntax tblSource :: Type

type Sql92TableSourceTableNameSyntax tblSource :: Type

Methods

tableNamed :: Sql92TableSourceTableNameSyntax tblSource -> tblSource

tableFromSubSelect :: Sql92TableSourceSelectSyntax tblSource -> tblSource

tableFromValues :: [[Sql92TableSourceExpressionSyntax tblSource]] -> tblSource

class IsSql92GroupingSyntax grouping where

Associated Types

type Sql92GroupingExpressionSyntax grouping :: Type

Methods

groupByExpressions :: [Sql92GroupingExpressionSyntax grouping] -> grouping

class (IsSql92TableSourceSyntax (Sql92FromTableSourceSyntax from), IsSql92ExpressionSyntax (Sql92FromExpressionSyntax from)) => IsSql92FromSyntax from where

Associated Types

type Sql92FromTableSourceSyntax from :: Type

type Sql92FromExpressionSyntax from :: Type

Methods

fromTable :: Sql92FromTableSourceSyntax from -> Maybe (Text, Maybe [Text]) -> from

innerJoin :: from -> from -> Maybe (Sql92FromExpressionSyntax from) -> from

leftJoin :: from -> from -> Maybe (Sql92FromExpressionSyntax from) -> from

rightJoin :: from -> from -> Maybe (Sql92FromExpressionSyntax from) -> from

class IsSql92FromSyntax from => IsSql92FromOuterJoinSyntax from where

Methods

outerJoin :: from -> from -> Maybe (Sql92FromExpressionSyntax from) -> from