Database.Beam.Backend.Types

Documentation

class BeamBackend be

Class for all Beam backends

Associated Types

type BackendFromField be :: Type -> Constraint

Requirements to marshal a certain type from a database of a particular backend

data Exposed x

newtype mainly used to inspect the tag structure of a particular Beamable. Prevents overlapping instances in some case. Usually not used in end-user code.

data Nullable (c :: Type -> Type) x

Support for NULLable Foreign Key references.

data MyTable f = MyTable
               { nullableRef :: PrimaryKey AnotherTable (Nullable f)
               , ... }
                deriving (Generic, Typeable)

See Columnar for more information.