Database.Beam.Query.Adhoc

Description

This module contains support for defining "ad-hoc" queries. That is queries on tables that do not necessarily have corresponding Beamable table types.

Documentation

class Adhoc structure where

Associated Types

type AdhocTable structure (f :: Type -> Type) :: Type

Methods

mkAdhocField :: (forall a. Text -> f a) -> structure -> AdhocTable structure f

data NamedField a

table_ :: forall be db structure s. (Adhoc structure, BeamSqlBackend be, Projectible be (AdhocTable structure (QExpr be s))) => Maybe Text -> Text -> structure -> Q be db s (AdhocTable structure (QExpr be s))

Introduce a table into a query without using the Beamable and Database machinery.

The first argument is the optional name of the schema the table is in and the second is the name of the table to source from.

The third argument is a tuple (or any nesting of tuples) where each value is of type NamedField (use field_ to construct).

The return value is a tuple (or any nesting of tuples) of the same shape as structure but where each value is a QExpr.

For example, to source from the table Table1, with fields Field1 (A boolean), Field2 (a timestamp), and Field3 (a string)

table_ Nothing "Table1" ( field_ @Bool "Field1", field_ @UTCTime "Field2", field_ @Text "Field3" )

field_ :: forall a. Text -> NamedField a

Used to construct NamedFields, most often with an explicitly applied type.

The type can be omitted if the value is used unambiguously elsewhere.