Database.Beam.Query.CTE

Documentation

data Recursiveness be where

Constructors

Nonrecursive :: Recursiveness be

Recursive :: IsSql99RecursiveCommonTableExpressionSelectSyntax (BeamSqlBackendSelectSyntax be) => Recursiveness be

newtype With be (db :: (Type -> Type) -> Type) a

Monad in which SELECT statements can be made (via selecting) and bound to result names for re-use later. This has the advantage of only computing each result once. In SQL, this is translated to a common table expression.

Once introduced, results can be re-used in future queries with reuse.

With is also a member of MonadFix for backends that support recursive CTEs. In this case, you can use mdo or rec notation (with RecursiveDo enabled) to bind result values (again, using reuse) even before they're introduced.

See further documentation here.

Constructors

With

Fields

runWith :: WriterT (Recursiveness be, [BeamSql99BackendCTESyntax be]) (State Int) a

data QAnyScope

data ReusableQ be db res where

Query results that have been introduced into a common table expression via selecting that can be used in future queries with reuse.

Constructors

ReusableQ :: Proxy res -> (forall s. Proxy s -> Q be db s (WithRewrittenThread QAnyScope s res)) -> ReusableQ be db res

reusableForCTE :: forall be res db. (ThreadRewritable QAnyScope res, Projectible be res, BeamSqlBackend be) => Text -> ReusableQ be db res

selecting :: forall res be db. (BeamSql99CommonTableExpressionBackend be, HasQBuilder be, Projectible be res, ThreadRewritable QAnyScope res) => Q be db QAnyScope res -> With be db (ReusableQ be db res)

Introduce the result of a query as a result in a common table expression. The returned value can be used in future queries by applying reuse.

reuse :: forall s be db res. ReusableQ be db res -> Q be db s (WithRewrittenThread QAnyScope s res)

Introduces the result of a previous selecting (a CTE) into a new query