Database.Beam.Query.DataTypes

Documentation

newtype DataType be a

A data type in a given IsSql92DataTypeSyntax which describes a SQL type mapping to the Haskell type a

Constructors

DataType (BeamSqlBackendCastTargetSyntax be)

cast_ :: BeamSqlBackend be => QGenExpr ctxt be s a -> DataType be b -> QGenExpr ctxt be s b

Cast a value to a specific data type, specified using DataType.

Note: this may fail at run-time if the cast is invalid for a particular value

Data types

int :: (BeamSqlBackend be, Integral a) => DataType be a

SQL92 INTEGER data type

smallint :: (BeamSqlBackend be, Integral a) => DataType be a

SQL92 SMALLINT data type

bigint :: (BeamSqlBackend be, BeamSqlT071Backend be, Integral a) => DataType be a

SQL2008 Optional BIGINT data type

binary :: (BeamSqlBackend be, BeamSqlT021Backend be) => Maybe Word -> DataType be Integer

SQL2003 Optional BINARY data type

varbinary :: (BeamSqlBackend be, BeamSqlT021Backend be) => Maybe Word -> DataType be Integer

SQL2003 Optional VARBINARY data type

date :: BeamSqlBackend be => DataType be Day

SQL92 DATE data type

char :: BeamSqlBackend be => Maybe Word -> DataType be Text

SQL92 CHAR data type

varchar :: BeamSqlBackend be => Maybe Word -> DataType be Text

SQL92 VARCHAR data type

nationalVarchar :: BeamSqlBackend be => Maybe Word -> DataType be Text

SQL92 NATIONAL CHARACTER VARYING data type

nationalChar :: BeamSqlBackend be => Maybe Word -> DataType be Text

SQL92 NATIONAL CHARACTER data type

double :: BeamSqlBackend be => DataType be Double

SQL92 DOUBLE data type

numeric :: BeamSqlBackend be => Maybe (Word, Maybe Word) -> DataType be Scientific

SQL92 NUMERIC data type

timestamptz :: BeamSqlBackend be => DataType be LocalTime

SQL92 TIMESTAMP WITH TIME ZONE data type

timestamp :: BeamSqlBackend be => DataType be LocalTime

SQL92 TIMESTAMP WITHOUT TIME ZONE data type

time :: BeamSqlBackend be => Maybe Word -> DataType be TimeOfDay

SQL92 TIME data type

boolean :: BeamSql99DataTypeBackend be => DataType be Bool

SQL99 BOOLEAN data type

characterLargeObject :: BeamSql99DataTypeBackend be => DataType be Text

SQL99 CLOB data type

binaryLargeObject :: BeamSql99DataTypeBackend be => DataType be Text

SQL99 BLOB data type

array :: (Typeable a, BeamSql99DataTypeBackend be) => DataType be a -> Int -> DataType be (Vector a)

SQL99 array data types

maybeType :: DataType be a -> DataType be (Maybe a)

Haskell requires DataTypes to match exactly. Use this function to convert a DataType that expects a concrete value to one expecting a Maybe