Database.Beam.Query.Internal

Documentation

type ProjectibleInBackend be a = (Projectible be a, ProjectibleValue be a)

type TablePrefix = Text

data QF be (db :: (Type -> Type) -> Type) s next where

Constructors

QDistinct :: Projectible be r => (r -> WithExprContext (BeamSqlBackendSetQuantifierSyntax be)) -> QM be db s r -> (r -> next) -> QF be db s next

QAll :: Projectible be r => (TablePrefix -> Text -> BeamSqlBackendFromSyntax be) -> (Text -> r) -> (r -> Maybe (WithExprContext (BeamSqlBackendExpressionSyntax be))) -> ((Text, r) -> next) -> QF be db s next

QArbitraryJoin :: Projectible be r => QM be db (QNested s) r -> Text -> (BeamSqlBackendFromSyntax be -> BeamSqlBackendFromSyntax be -> Maybe (BeamSqlBackendExpressionSyntax be) -> BeamSqlBackendFromSyntax be) -> (r -> Maybe (WithExprContext (BeamSqlBackendExpressionSyntax be))) -> (r -> next) -> QF be db s next

QTwoWayJoin :: (Projectible be a, Projectible be b) => QM be db (QNested s) a -> QM be db (QNested s) b -> (BeamSqlBackendFromSyntax be -> BeamSqlBackendFromSyntax be -> Maybe (BeamSqlBackendExpressionSyntax be) -> BeamSqlBackendFromSyntax be) -> ((a, b) -> Maybe (WithExprContext (BeamSqlBackendExpressionSyntax be))) -> ((a, b) -> next) -> QF be db s next

QSubSelect :: Projectible be r => QM be db (QNested s) r -> (r -> next) -> QF be db s next

QGuard :: WithExprContext (BeamSqlBackendExpressionSyntax be) -> next -> QF be db s next

QLimit :: Projectible be r => Integer -> QM be db (QNested s) r -> (r -> next) -> QF be db s next

QOffset :: Projectible be r => Integer -> QM be db (QNested s) r -> (r -> next) -> QF be db s next

QSetOp :: Projectible be r => (BeamSqlBackendSelectTableSyntax be -> BeamSqlBackendSelectTableSyntax be -> BeamSqlBackendSelectTableSyntax be) -> QM be db (QNested s) r -> QM be db (QNested s) r -> (r -> next) -> QF be db s next

QOrderBy :: Projectible be r => (r -> WithExprContext [BeamSqlBackendOrderingSyntax be]) -> QM be db (QNested s) r -> (r -> next) -> QF be db s next

QWindowOver :: (ProjectibleWithPredicate WindowFrameContext be (WithExprContext (BeamSqlBackendWindowFrameSyntax' be)) window, Projectible be r, Projectible be a) => (r -> window) -> (r -> window -> a) -> QM be db (QNested s) r -> (a -> next) -> QF be db s next

QAggregate :: (Projectible be grouping, Projectible be a) => (a -> TablePrefix -> (Maybe (BeamSqlBackendGroupingSyntax be), grouping)) -> QM be db (QNested s) a -> (grouping -> next) -> QF be db s next

QForceSelect :: Projectible be r => (r -> BeamSqlBackendSelectTableSyntax be -> [BeamSqlBackendOrderingSyntax be] -> Maybe Integer -> Maybe Integer -> BeamSqlBackendSelectSyntax be) -> QM be db (QNested s) r -> (r -> next) -> QF be db s next

type QM be db s = F (QF be db s)

newtype Q be (db :: (Type -> Type) -> Type) s a

The type of queries over the database db returning results of type a. The s argument is a threading argument meant to restrict cross-usage of QExprs. syntax represents the SQL syntax that this query is building.

Constructors

Q

Fields

runQ :: QM be db s a

data QInternal

data QNested s

data QField s ty

Constructors

QField

Fields

qFieldShouldQualify :: !Bool

qFieldTblName :: !Text

qFieldName :: !Text

newtype QAssignment be s

Constructors

QAssignment

Fields

unQAssignment :: [(BeamSqlBackendFieldNameSyntax be, BeamSqlBackendExpressionSyntax be)]

newtype QFieldAssignment be tbl a

Constructors

QFieldAssignment (forall s. tbl (QExpr be s) -> Maybe (QExpr be s a))

QGenExpr type

data QAggregateContext

data QGroupingContext

data QValueContext

data QWindowingContext

data QWindowFrameContext

newtype QGenExpr context be s t

The type of lifted beam expressions that will yield the haskell type t.

context is a type-level representation of the types of expressions this can contain. For example, QAggregateContext represents expressions that may contain aggregates, and QWindowingContext represents expressions that may contain OVER.

syntax is the expression syntax being built (usually a type that implements IsSql92ExpressionSyntax at least, but not always).

s is a state threading parameter that prevents QExprs from incompatible sources to be combined. For example, this is used to prevent monadic joins from depending on the result of previous joins (so-called LATERAL joins).

Constructors

QExpr (TablePrefix -> BeamSqlBackendExpressionSyntax be)

newtype QOrd be s t

Constructors

QOrd (TablePrefix -> BeamSqlBackendOrderingSyntax be)

type WithExprContext a = TablePrefix -> a

type QExpr = QGenExpr QValueContext

QExprs represent expressions not containing aggregates.

type QAgg = QGenExpr QAggregateContext

type QWindowExpr = QGenExpr QWindowingContext

type QGroupExpr = QGenExpr QGroupingContext

newtype QWindow be s

Constructors

QWindow (WithExprContext (BeamSqlBackendWindowFrameSyntax be))

newtype QFrameBounds be

Constructors

QFrameBounds (Maybe (BeamSqlBackendWindowFrameBoundsSyntax be))

newtype QFrameBound be

Constructors

QFrameBound (BeamSqlBackendWindowFrameBoundSyntax be)

qBinOpE :: BeamSqlBackend be => (BeamSqlBackendExpressionSyntax be -> BeamSqlBackendExpressionSyntax be -> BeamSqlBackendExpressionSyntax be) -> QGenExpr context be s a -> QGenExpr context be s b -> QGenExpr context be s c

unsafeRetype :: QGenExpr ctxt be s a -> QGenExpr ctxt be s a'

Sql Projections

class Typeable context => AggregateContext context

Typeclass for all haskell data types that can be used to create a projection in a SQL select statement. This includes all tables as well as all tuple classes. Projections are only defined on tuples up to size 5. If you need more, follow the implementations here.

type family ContextName a :: Symbol

type family IsAggregateContext a :: Constraint where ...

Equations

IsAggregateContext QAggregateContext = ()

IsAggregateContext QGroupingContext = ()

IsAggregateContext a = TypeError (('Text "Non-aggregate expression where aggregate expected." :$$: (('Text "Got " :<>: 'Text (ContextName a)) :<>: 'Text ". Expected an aggregate or a grouping")) :$$: AggregateContextSuggestion a)

type family AggregateContextSuggestion a :: ErrorMessage where ...

Equations

AggregateContextSuggestion QValueContext = 'Text "Perhaps you forgot to wrap a value expression with 'group_'"

AggregateContextSuggestion QWindowingContext = 'Text "Perhaps you meant to use 'window_' instead of 'aggregate_'"

AggregateContextSuggestion b = 'Text ""

class Typeable context => ValueContext context

class Typeable context => WindowFrameContext context

type family IsWindowFrameContext a :: Constraint where ...

Equations

IsWindowFrameContext QWindowFrameContext = ()

IsWindowFrameContext a = TypeError ('Text "Expected window frame." :$$: (('Text "Got " :<>: 'Text (ContextName a)) :<>: 'Text ". Expected a window frame"))

class AnyType a

type family IsValueContext a :: Constraint where ...

Equations

IsValueContext QValueContext = ()

IsValueContext a = TypeError (('Text "Non-scalar context in projection" :$$: (('Text "Got " :<>: 'Text (ContextName a)) :<>: 'Text ". Expected a value")) :$$: ValueContextSuggestion a)

type family ValueContextSuggestion a :: ErrorMessage where ...

Equations

ValueContextSuggestion QWindowingContext = 'Text "Use 'window_' to projecct aggregate expressions to the value level"

ValueContextSuggestion QAggregateContext = 'Text "Aggregate functions and groupings cannot be contained in value expressions." :$$: 'Text "Use 'aggregate_' to compute aggregations at the value level."

ValueContextSuggestion QGroupingContext = ValueContextSuggestion QAggregateContext

ValueContextSuggestion _ = 'Text ""

type Projectible be = ProjectibleWithPredicate AnyType be (WithExprContext (BeamSqlBackendExpressionSyntax' be))

type ProjectibleValue be = ProjectibleWithPredicate ValueContext be (WithExprContext (BeamSqlBackendExpressionSyntax' be))

class ThreadRewritable (s :: Type) (a :: Type) | a -> s where

Associated Types

type WithRewrittenThread s (s' :: Type) a :: Type

Methods

rewriteThread :: Proxy s' -> a -> WithRewrittenThread s s' a

class ContextRewritable a where

Associated Types

type WithRewrittenContext a ctxt :: Type

Methods

rewriteContext :: Proxy ctxt -> a -> WithRewrittenContext a ctxt

newtype BeamSqlBackendExpressionSyntax' be

Constructors

BeamSqlBackendExpressionSyntax'

Fields

fromBeamSqlBackendExpressionSyntax :: BeamSqlBackendExpressionSyntax be

newtype BeamSqlBackendWindowFrameSyntax' be

Constructors

BeamSqlBackendWindowFrameSyntax'

Fields

fromBeamSqlBackendWindowFrameSyntax :: BeamSqlBackendWindowFrameSyntax be

class ProjectibleWithPredicate (contextPredicate :: Type -> Constraint) be res a | a -> be where

Methods

project' :: Monad m => Proxy contextPredicate -> Proxy (be, res) -> (forall context. contextPredicate context => Proxy context -> Proxy be -> res -> m res) -> a -> m a

projectSkeleton' :: Monad m => Proxy contextPredicate -> Proxy (be, res) -> (forall context. contextPredicate context => Proxy context -> Proxy be -> m res) -> m a

project :: forall be a. Projectible be a => Proxy be -> a -> WithExprContext [BeamSqlBackendExpressionSyntax be]

reproject :: forall be a. (BeamSqlBackend be, Projectible be a) => Proxy be -> (Int -> BeamSqlBackendExpressionSyntax be) -> a -> a

tableFieldsToExpressions :: (BeamSqlBackend be, Beamable table) => TableSettings table -> Text -> table (QGenExpr ctxt be s)

suitable as argument to QAll in the case of a table result

mkFieldsSkeleton :: forall be res m. (Projectible be res, MonadState Int m) => (Int -> m (WithExprContext (BeamSqlBackendExpressionSyntax' be))) -> m res

mkFieldNames :: forall be res. (BeamSqlBackend be, Projectible be res) => (Text -> BeamSqlBackendFieldNameSyntax be) -> (res, [Text])

tableNameFromEntity :: IsSql92TableNameSyntax name => DatabaseEntityDescriptor be (TableEntity tbl) -> name

rescopeQ :: QM be db s res -> QM be db s' res