Database.Beam.Query.Types

Documentation

data Q be (db :: (Type -> Type) -> Type) s a

The type of queries over the database db returning results of type a. The s argument is a threading argument meant to restrict cross-usage of QExprs. syntax represents the SQL syntax that this query is building.

type QExpr = QGenExpr QValueContext

QExprs represent expressions not containing aggregates.

newtype QGenExpr context be s t

The type of lifted beam expressions that will yield the haskell type t.

context is a type-level representation of the types of expressions this can contain. For example, QAggregateContext represents expressions that may contain aggregates, and QWindowingContext represents expressions that may contain OVER.

syntax is the expression syntax being built (usually a type that implements IsSql92ExpressionSyntax at least, but not always).

s is a state threading parameter that prevents QExprs from incompatible sources to be combined. For example, this is used to prevent monadic joins from depending on the result of previous joins (so-called LATERAL joins).

Constructors

QExpr (TablePrefix -> BeamSqlBackendExpressionSyntax be)

type family QExprToIdentity x

type family QExprToField x

data QWindow be s

type Projectible be = ProjectibleWithPredicate AnyType be (WithExprContext (BeamSqlBackendExpressionSyntax' be))

class BeamSqlBackend be => HasQBuilder be where

Methods

buildSqlQuery :: Projectible be a => TablePrefix -> Q be db s a -> BeamSqlBackendSelectSyntax be